requirements.txt
streamlit
torch
torchvision
timm
scikit-learn
numpy
opencv-python-headless
Pillow